requirements.txt
APScheduler==3.6.0
authy==2.2.3
autopep8==1.4.4
awsebcli==3.18.2
boto3==1.14.23
botocore==1.15.49
cement==2.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==2.2.1
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-environ==0.4.5
django-heroku==0.3.1
django-storages==1.9.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
docutils==0.15.2
future==0.16.0
gunicorn==20.0.4
idna==2.7
jmespath==0.10.0
pathspec==0.5.9
paystack==1.5.0
paystackpy @ file:///C:/Users/User/paystack/paystackpy
Pillow==6.2.1
psycopg2==2.8.5
pycodestyle==2.5.0
PyJWT==1.7.1
pypiwin32==223
python-dateutil==2.8.0
pytz==2019.3
pywin32==228
PyYAML==5.3.1
requests==2.20.1
s3transfer==0.3.3
semantic-version==2.5.0
simplejson==3.17.0
six==1.11.0
SQLAlchemy==1.3.19
sqlparse==0.3.0
termcolor==1.1.0
tzlocal==2.0.0
urllib3==1.24.3
wcwidth==0.1.9
whitenoise==5.1.0
